+++ src/phenoml/construe/errors/__init__.py
@@ -5,6 +5,7 @@
 from .bad_request_error import BadRequestError
 from .conflict_error import ConflictError
 from .failed_dependency_error import FailedDependencyError
+from .forbidden_error import ForbiddenError
 from .internal_server_error import InternalServerError
 from .not_found_error import NotFoundError
 from .not_implemented_error import NotImplementedError
@@ -15,6 +16,7 @@ __all__ = [
     "BadRequestError",
     "ConflictError",
     "FailedDependencyError",
+    "ForbiddenError",
     "InternalServerError",
     "NotFoundError",
     "NotImplementedError",