+++ src/phenoml/construe/types/__init__.py
@@ -7,6 +7,7 @@ from .code_response import CodeResponse
 from .code_system_details import CodeSystemDetails
 from .code_system_info import CodeSystemInfo
 from .construe_upload_code_system_response import ConstrueUploadCodeSystemResponse
+from .delete_code_system_response import DeleteCodeSystemResponse
 from .extract_codes_result import ExtractCodesResult
 from .extract_request_config import ExtractRequestConfig
 from .extract_request_config_chunking_method import ExtractRequestConfigChunkingMethod
@@ -14,6 +15,7 @@ from .extract_request_config_validation_method import ExtractRequestConfigValida
 from .extract_request_system import ExtractRequestSystem
 from .extracted_code_result import ExtractedCodeResult
 from .get_code_response import GetCodeResponse
+from .get_code_system_detail_response import GetCodeSystemDetailResponse
 from .list_code_systems_response import ListCodeSystemsResponse
 from .list_codes_response import ListCodesResponse
 from .semantic_search_response import SemanticSearchResponse
@@ -28,6 +30,7 @@ __all__ = [
     "CodeSystemDetails",
     "CodeSystemInfo",
     "ConstrueUploadCodeSystemResponse",
+    "DeleteCodeSystemResponse",
     "ExtractCodesResult",
     "ExtractRequestConfig",
     "ExtractRequestConfigChunkingMethod",
@@ -35,6 +38,7 @@ __all__ = [
     "ExtractRequestSystem",
     "ExtractedCodeResult",
     "GetCodeResponse",
+    "GetCodeSystemDetailResponse",
     "ListCodeSystemsResponse",
     "ListCodesResponse",
     "SemanticSearchResponse",